# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Saukele Digital Registry API
  version: 1.0.0
  description: |
    REST API for **Saukele** — a digital wedding registry that models Kazakh
    family-tree kinship, multi-currency pool funding with snapshotted exchange
    rates, and traditional-gift logistics.

    ## Conventions
    - All timestamps are ISO-8601 UTC.
    - All amounts are decimals serialized as JSON numbers; the canonical currency
      is KZT. Foreign currencies are stored alongside the rate that was active at
      transaction time and are NEVER recalculated.
    - All endpoints (except `/auth/register`, `/auth/login`, `/auth/refresh`,
      `/healthz`) require a valid JWT access token.
    - Pagination on list endpoints is **cursor-based** (keyset). The cursor is
      opaque to the client — pass back exactly what `next_cursor` returned.
    - Errors follow a single shape (`ErrorResponse`).

  contact:
    name: Saukele Backend Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.saukele.kz/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      description: Opaque keyset cursor returned by the previous page response.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    GuestIdPath:
      name: guest_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code, status_code]
      properties:
        error:
          type: string
          description: Human-readable summary.
          example: "Validation failed"
        code:
          type: string
          description: Stable machine-readable code (UPPER_SNAKE).
          example: "VALIDATION_ERROR"
        status_code:
          type: integer
          example: 422
        details:
          type: array
          description: Field-level errors when applicable.
          items:
            type: object
            properties:
              field: { type: string, example: "email" }
              issue: { type: string, example: "must be a valid email" }
        request_id:
          type: string
          example: "req_01HXYZ..."
    Pagination:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: string
          nullable: true
          description: Pass to `?cursor=` for the next page. `null` = last page.
    Role:
      type: string
      enum: [HOST, GUEST, VENDOR, COURIER, ADMIN]

    # ── Auth ──────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, username, password, role]
      properties:
        email:    { type: string, format: email, example: "aizhan@example.kz" }
        username: { type: string, minLength: 3, maxLength: 32, example: "aizhan" }
        password:
          type: string
          minLength: 10
          description: Min 10 chars, must contain at least one letter and one digit.
          example: "Saukele2025!"
        role:
          $ref: '#/components/schemas/Role'
          description: ADMIN cannot self-register.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string }
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:  { type: string, example: "eyJhbGciOi..." }
        refresh_token: { type: string, example: "eyJhbGciOi..." }
        token_type:    { type: string, example: "Bearer" }
        expires_in:    { type: integer, description: "access TTL in seconds", example: 900 }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    User:
      type: object
      properties:
        id:        { type: string, format: uuid }
        email:     { type: string, format: email }
        username:  { type: string }
        role:      { $ref: '#/components/schemas/Role' }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }

    # ── Registry ──────────────────────────────────────────────────────────────
...
    FamilyTreeNode:
      type: object
      description: One node in the recursive family-tree response.
      properties:
        id:            { type: string, format: uuid }
        parent_id:     { type: string, format: uuid, nullable: true }
        display_name:  { type: string }
        kinship_label: { type: string }
        tier_rank:     { type: integer }
        depth:         { type: integer, description: "0 = requested root" }
        suggested_kzt: { type: number, description: "tier obligation amount" }
        funded_kzt:    { type: number, description: "already contributed by this guest" }

    # ── Contribution ──────────────────────────────────────────────────────────
...
  responses:
...
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated, but role or tier does not permit this action.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Unprocessable:
      description: Validation passed type-wise, but business rule rejected the request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Creates a new user with the requested role (ADMIN is rejected).
        Rate-limited to 5 attempts per minute per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
            examples:
              host:
                value:
                  email: "aizhan@example.kz"
                  username: "aizhan"
                  password: "Saukele2025!"
                  role: "HOST"
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '409':
          description: Email or username already taken.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive an access + refresh token pair
      description: Rate-limited to 5 attempts per minute per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Token pair issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: New token pair (refresh rotated).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: Refresh token invalid, expired, or revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the caller's refresh token
      description: |
        Marks the refresh token as revoked in the database AND adds it to the
        Redis denylist for instant invalidation.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '204': { description: "Logged out." }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Users]
      summary: Get the authenticated user's profile
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ── Registries ─────────────────────────────────────────────────────────────
...
  /guests/{guest_id}/family-tree:
    parameters:
      - $ref: '#/components/parameters/GuestIdPath'
    get:
      tags: [FamilyTree]
      summary: Recursive family-tree subtree with gift-tier obligations
      description: |
        **Assignment-mandated recursive endpoint.** Implemented via PostgreSQL
        `WITH RECURSIVE` over the self-referential `guests.parent_id` column.
        Each node carries:
          - `depth` (0 at the requested root),
          - `suggested_kzt` — the per-tier obligation amount,
          - `funded_kzt`   — what this guest has already contributed.

        Cycle-guarded and depth-capped at 10.
      parameters:
        - name: max_depth
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 10, default: 10 }
      responses:
        '200':
          description: Flat list of nodes (DFS order, depth ascending).
          content:
            application/json:
              schema:
                type: object
                properties:
                  root_id: { type: string, format: uuid }
                  nodes:
                    type: array
                    items: { $ref: '#/components/schemas/FamilyTreeNode' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Admin ──────────────────────────────────────────────────────────────────
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users (ADMIN only)
      description: RBAC test surface — non-admin tokens MUST receive 403.
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Page of users.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }